优化代码 - Lesyuex

+++ src/main/java/com/jobeth/vo/StockKLineVo.java
@@ -62,5 +62,5 @@ public class StockKLineVo {
      * 成交量
      */
     @QtIndex("8")
-    private Integer volume;
+    private Double volume;
 }
